long long storage

+++ src/books.h
@@ -75,7 +75,7 @@ public:
   Author author{};
   KeyWord key_word{};
   double price = 0;
-  int storage = 0;
+  long long storage = 0;
 public:
   Book() = default;
